# Improve GPU performance by changing the offset computation of the `FullGridCellList`