Based on #49.

I basically changed what the commit says: "Subtract `min_corner` in `cell_coords` instead of subtracting `min_cell` in `getindex`".

Previously, `cell_coords` returned the absolute cell coordinates, then the neighbor cells around these were computed, and for each of those neighbor cells, the offset `min_cell` had to be subtracted.
Now, I subtract the `min_corner` in `cell_coordinates`, so that we don't have to do this for every neighboring cell.

On the CPU, the difference is small:
```
julia> plot_benchmarks(benchmark_wcsph, (118, 118, 118), 3)
new version
with 118x118x118 = 1643032 particles finished in 243.722 ms

first version
with 118x118x118 = 1643032 particles finished in 262.962 ms

new version
with 187x187x187 = 6539203 particles finished in 1.003 s

first version
with 187x187x187 = 6539203 particles finished in 1.062 s

new version
with 297x297x297 = 26198073 particles finished in 4.070 s

first version
with 297x297x297 = 26198073 particles finished in 4.284 s

julia> plot_benchmarks(benchmark_n_body, (118, 118, 118), 3)
new version
with 118x118x118 = 1643032 particles finished in 103.480 ms

first version
with 118x118x118 = 1643032 particles finished in 116.125 ms

new version
with 187x187x187 = 6539203 particles finished in 424.271 ms

first version
with 187x187x187 = 6539203 particles finished in 471.688 ms

new version
with 297x297x297 = 26198073 particles finished in 1.709 s

first version
with 297x297x297 = 26198073 particles finished in 1.897 s
```
On the GPU (RTX 3090), however, the difference is huge. Even for the real-life WCSPH benchmark:
```
julia> plot_benchmarks(benchmark_wcsph_gpu, (118, 118, 118), 3)
new version
with 118x118x118 = 1643032 particles finished in 252.205 ms

first version
with 118x118x118 = 1643032 particles finished in 405.049 ms

new version
with 187x187x187 = 6539203 particles finished in 1.003 s

first version
with 187x187x187 = 6539203 particles finished in 1.643 s

new version
with 297x297x297 = 26198073 particles finished in 3.972 s

first version
with 297x297x297 = 26198073 particles finished in 6.602 s

julia> plot_benchmarks(benchmark_n_body_gpu, (118, 118, 118), 3)
new version
with 118x118x118 = 1643032 particles finished in 85.513 ms

first version
with 118x118x118 = 1643032 particles finished in 135.293 ms

new version
with 187x187x187 = 6539203 particles finished in 338.926 ms

first version
with 187x187x187 = 6539203 particles finished in 536.936 ms

new version
with 297x297x297 = 26198073 particles finished in 1.352 s

first version
with 297x297x297 = 26198073 particles finished in 2.177 s
```